requirements.txt
feedparser>=6.0
anthropic>=0.40
apscheduler>=3.10
python-dotenv>=1.0
jinja2>=3.1
requests>=2.31
python-dateutil>=2.9
